02-SRC/DreamerV3/requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0

# Hockey environment
hockey-env>=1.0.0

# Logging and visualization
wandb>=0.15.0
tqdm>=4.65.0
imageio>=2.31.0

# Optional: External DreamerV3 implementations for Phase 1 testing
# Uncomment ONE of the following to use an existing implementation:
#
# Option 1: dreamer-pytorch (recommended for PyTorch users)
# pip install dreamer-pytorch
#
# Option 2: sheeprl (comprehensive RL library with DreamerV3)
# pip install sheeprl
#
# Option 3: Use our custom implementation (Phase 2 - no additional deps needed)

# Development
pytest>=7.0.0
